// frontend/src/openapi/generated/model/tag.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Zete
 * 메모 서비스
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface Tag
 */
export interface Tag {
    /**
     * 
     * @type {number}
     * @memberof Tag
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof Tag
     */
    'name'?: string;
    /**
     * 
     * @type {number}
     * @memberof Tag
     */
    'userId'?: number;
    /**
     * 
     * @type {number}
     * @memberof Tag
     */
    'memoId'?: number;
    /**
     * 
     * @type {number}
     * @memberof Tag
     */
    'cateId'?: number | null;
}